Update, possible solution to Probelm 1673.
using a min heap to order the values, I have this concept of a solution
window. for every new valid number to add to solution vector, I move
lower bound to new index and increment upper bound by one, which ensures
that the next valid value is a subsequent one.